+++ src/state/metadata/constants.ts
@@ -13,18 +13,9 @@ export const RECEIVE_SIMULARIUM_FILE = makeMetadataConstant(
 export const LOAD_LOCAL_FILE_IN_VIEWER = makeMetadataConstant(
     "load-local-file"
 );
-export const SET_SIMULARIUM_CONTROLLER = makeMetadataConstant(
-    "set-sim-controller"
-);
-export const SET_VIEWER_STATUS = makeMetadataConstant("set-viewer-status");
 export const LOAD_NETWORKED_FILE_IN_VIEWER = makeMetadataConstant(
     "load-networked-file"
 );
 export const REQUEST_PLOT_DATA = makeMetadataConstant("request-plot-data");
 export const CLEAR_SIMULARIUM_FILE = makeMetadataConstant("clear-sim-file");
 export const LOAD_FILE_VIA_URL = makeMetadataConstant("load-file-via-url");
-
-export const VIEWER_EMPTY = "empty";
-export const VIEWER_LOADING = "loading";
-export const VIEWER_ERROR = "error";
-export const VIEWER_SUCCESS = "success";
