+++ src/state/metadata/reducer.ts
@@ -8,17 +8,12 @@ import {
     RECEIVE_AGENT_IDS,
     RECEIVE_AGENT_NAMES,
     RECEIVE_SIMULARIUM_FILE,
-    SET_SIMULARIUM_CONTROLLER,
-    SET_VIEWER_STATUS,
-    VIEWER_ERROR,
-    VIEWER_EMPTY,
     CLEAR_SIMULARIUM_FILE,
 } from "./constants";
 import {
     MetadataStateBranch,
     ReceiveAction,
     ClearSimFileDataAction,
-    SetViewerStatusAction,
 } from "./types";
 
 export const initialState = {
@@ -34,9 +29,6 @@ export const initialState = {
         data: null,
         lastModified: null,
     },
-    simulariumController: null,
-    viewerStatus: VIEWER_EMPTY,
-    viewerError: null,
 };
 
 const actionToConfigMap: TypeToDescriptionMap = {
@@ -78,33 +70,6 @@ const actionToConfigMap: TypeToDescriptionMap = {
         perform: (state: MetadataStateBranch) => ({
             ...state,
             simulariumFile: initialState.simulariumFile,
-        }),
-    },
-    [SET_SIMULARIUM_CONTROLLER]: {
-        accepts: (action: AnyAction): action is ReceiveAction =>
-            action.type === SET_SIMULARIUM_CONTROLLER,
-        perform: (state: MetadataStateBranch, action: ReceiveAction) => ({
-            ...state,
-            simulariumController: action.payload,
-        }),
-    },
-    [SET_VIEWER_STATUS]: {
-        accepts: (action: AnyAction): action is SetViewerStatusAction =>
-            action.type === SET_VIEWER_STATUS,
-        perform: (
-            state: MetadataStateBranch,
-            action: SetViewerStatusAction
-        ) => ({
-            ...state,
-            viewerStatus: action.payload.status,
-            viewerError:
-                action.payload.status === VIEWER_ERROR
-                    ? {
-                          message: action.payload.errorMessage,
-                          htmlData: action.payload.htmlData,
-                          onClose: action.payload.onClose,
-                      }
-                    : "",
         }),
     },
 };
