+++ src/state/simularium/actions.ts
@@ -1,0 +1,9 @@
+import { RECEIVE_METADATA } from "./constants";
+import { MetadataStateBranch } from "./types";
+
+export function receiveMetadata(payload: MetadataStateBranch): ReceiveAction {
+    return {
+        payload,
+        type: RECEIVE_METADATA,
+    };
+}
