+++ src/state/simularium/constants.ts
@@ -1,0 +1,8 @@
+import { makeConstant } from "../util";
+const BRANCH_NAME = "simularium";
+const makeSimulariumConstant = (constant: string) =>
+    makeConstant(BRANCH_NAME, constant);
+
+export const SET_SIMULARIUM_CONTROLLER = makeSimulariumConstant(
+    "set-sim-controller"
+);
