+++ src/state/simularium/types.ts
@@ -1,0 +1,8 @@
+export interface SimulariumStateBranch {
+    [key: string]: any;
+}
+
+export interface ReceiveAction {
+    payload: SimulariumStateBranch;
+    type: string;
+}
